+++ CHANGELOG.md
@@ -6,11 +6,14 @@
 
 - ** Destination connector adding**: MongoDB
 
+- List all the tools in a table.
+
 - Capability to log calls to UnstructuredClient
 
 ### Fixes
 
 - avoid client startup error if env not defined
+- force server using env var from .env instead of fetching them from system first
 
 
 ## 0.1.1