+++ CHANGELOG.md
@@ -8,9 +8,12 @@
 
 - ** Destination connector adding**: MongoDB, Databricks Volumes, Databricks Volumes Delta Table
 
+- List all the tools in a table.
+
 ### Fixes
 
 - avoid client startup error if env not defined
+- force server using env var from .env instead of fetching them from system first
 
 
 ## 0.1.1