+++ CHANGELOG.md
@@ -6,11 +6,14 @@
 
 - ** Destination connector adding**: MongoDB
 
-- Capability to log request parameters to UnstructuredClient's AsyncHttpClient, no error response is logged by this mechanism.
+- List all the tools in a table.
+
+- Capability to only log request parameters to UnstructuredClient's AsyncHttpClient, no error response is logged by this mechanism.
 
 ### Fixes
 
 - avoid client startup error if env not defined
+- force server using env var from .env instead of fetching them from system first
 
 
 ## 0.1.1